+++ contrib/activity/definition/src/main/scala/com/krux/hyperion/activity/SftpActivity.scala
@@ -3,7 +3,6 @@ package com.krux.hyperion.activity
 import com.krux.hyperion.adt._
 import com.krux.hyperion.aws.AdpShellCommandActivity
 import com.krux.hyperion.datanode.S3DataNode
-import com.krux.hyperion.expression.ConstantExpression._
 import com.krux.hyperion.expression.{ Format, EncryptedParameter }
 import com.krux.hyperion.resource.Ec2Resource